+++ cider-interaction.el
@@ -1241,8 +1241,8 @@ currently selected buffer."
               '(t always only-in-repl)
             '(t always except-in-repl)))))
 
-(defun cider-new-error-buffer ()
-  "Return an empty error buffer, possibly displaying and/or selecting it.
+(defun cider-new-error-buffer (&optional mode)
+  "Return an empty error buffer using MODE.
 
 When deciding whether to display the buffer, takes into account both the
 value of `cider-show-error-buffer' and the currently selected buffer.
@@ -1250,8 +1250,8 @@ value of `cider-show-error-buffer' and the currently selected buffer.
 When deciding whether to select the buffer, takes into account the value of
 `cider-auto-select-error-buffer'."
   (if (cider--show-error-buffer-p)
-      (cider-popup-buffer cider-error-buffer cider-auto-select-error-buffer)
-    (cider-make-popup-buffer cider-error-buffer)))
+      (cider-popup-buffer cider-error-buffer cider-auto-select-error-buffer mode)
+    (cider-make-popup-buffer cider-error-buffer mode)))
 
 (defun cider--handle-err-eval-response (response)
   "Render eval RESPONSE into a new error buffer.
@@ -1275,7 +1275,7 @@ Uses the value of the `out' slot in RESPONSE."
 (defun cider--render-stacktrace-causes (causes)
   "If CAUSES is non-nil, render its contents into a new error buffer."
   (when causes
-    (let ((error-buffer (cider-new-error-buffer)))
+    (let ((error-buffer (cider-new-error-buffer #'cider-stacktrace-mode)))
       (cider-stacktrace-render error-buffer (reverse causes)))))
 
 (defun cider--handle-stacktrace-response (response causes)